pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "qjson-agents"
version = "0.1.0"
description = "QJSON Agents: recursive, memoryful local agents over Ollama"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Yan Desbiens"}]
dependencies = ["googlesearch-python", "requests"]

[project.scripts]
qjson-agents = "qjson_agents.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["qjson_agents*"]
